+++ crates/controller/src/api/signaling/ws/runner.rs
@@ -7,7 +7,7 @@ use super::modules::{
     AnyStream, DynBroadcastEvent, DynEventCtx, DynTargetedEvent, Modules, NoSuchModuleError,
 };
 use super::{
-    DestroyContext, Namespaced, NamespacedOutgoing, RabbitMqBinding, RabbitMqExchange,
+    DestroyContext, NamespacedCommand, NamespacedEvent, RabbitMqBinding, RabbitMqExchange,
     RabbitMqPublish, Timestamp,
 };
 use crate::api;
@@ -699,11 +699,11 @@ impl Runner {
                                 self.room_id.room_id(),
                             )),
                             control::rabbitmq::room_all_routing_key(),
-                            Namespaced {
+                            serde_json::to_string(&NamespacedCommand {
                                 namespace: moderation::NAMESPACE,
                                 payload: moderation::rabbitmq::Message::LeftWaitingRoom(self.id),
-                            }
-                            .to_json(),
+                            })
+                            .expect("Failed to convert namespaced to json"),
                         )
                         .await;
                     }
@@ -900,7 +900,7 @@ impl Runner {
     async fn handle_ws_message(&mut self, message: Message) {
         log::trace!("Received websocket message {:?}", message);
 
-        let value: Result<Namespaced<'_, Value>, _> = match message {
+        let value: Result<NamespacedCommand<'_, Value>, _> = match message {
             Message::Text(ref text) => serde_json::from_str(text),
             Message::Binary(ref binary) => serde_json::from_slice(binary),
             _ => unreachable!(),
@@ -1062,11 +1062,11 @@ impl Runner {
                                     .as_str(),
                             ),
                             control::rabbitmq::room_all_routing_key(),
-                            Namespaced {
+                            serde_json::to_string(&NamespacedCommand {
                                 namespace: moderation::NAMESPACE,
                                 payload: moderation::rabbitmq::Message::LeftWaitingRoom(self.id),
-                            }
-                            .to_json(),
+                            })
+                            .expect("Failed to convert namespaced to json"),
                         )
                         .await;
 
@@ -1287,12 +1287,12 @@ impl Runner {
 
         self.ws
             .send(Message::Text(
-                NamespacedOutgoing {
+                serde_json::to_string(&NamespacedEvent {
                     namespace: moderation::NAMESPACE,
                     timestamp,
                     payload: moderation::outgoing::Message::InWaitingRoom,
-                }
-                .to_json(),
+                })?
+                .into(),
             ))
             .await;
 
@@ -1300,11 +1300,11 @@ impl Runner {
             timestamp,
             Some(breakout::rabbitmq::global_exchange_name(self.room_id.room_id()).as_str()),
             control::rabbitmq::room_all_routing_key(),
-            Namespaced {
+            serde_json::to_string(&NamespacedCommand {
                 namespace: moderation::NAMESPACE,
                 payload: moderation::rabbitmq::Message::JoinedWaitingRoom(self.id),
-            }
-            .to_json(),
+            })
+            .expect("Failed to convert namespaced to json"),
         )
         .await;
 
@@ -1580,13 +1580,14 @@ impl Runner {
             }
             ).into();
 
-            let namespaced = match serde_json::from_slice::<Namespaced<Value>>(&delivery.data) {
-                Ok(namespaced) => namespaced,
-                Err(e) => {
-                    log::error!("Failed to read incoming rabbit-mq message, {}", e);
-                    return;
-                }
-            };
+            let namespaced =
+                match serde_json::from_slice::<NamespacedCommand<Value>>(&delivery.data) {
+                    Ok(namespaced) => namespaced,
+                    Err(e) => {
+                        log::error!("Failed to read incoming rabbit-mq message, {}", e);
+                        return;
+                    }
+                };
 
             if namespaced.namespace == NAMESPACE {
                 let msg = match serde_json::from_value::<rabbitmq::Message>(namespaced.payload) {
@@ -1735,12 +1736,12 @@ impl Runner {
 
                         self.ws
                             .send(Message::Text(
-                                NamespacedOutgoing {
+                                serde_json::to_string(&NamespacedEvent {
                                     namespace: moderation::NAMESPACE,
                                     timestamp,
                                     payload: moderation::outgoing::Message::Accepted,
-                                }
-                                .to_json(),
+                                })?
+                                .into(),
                             ))
                             .await;
                     }
@@ -1803,26 +1804,26 @@ impl Runner {
 
                 self.ws
                     .send(Message::Text(
-                        NamespacedOutgoing {
+                        serde_json::to_string(&NamespacedEvent {
                             namespace: moderation::NAMESPACE,
                             timestamp,
                             payload: moderation::outgoing::Message::RaisedHandResetByModerator {
                                 issued_by,
                             },
-                        }
-                        .to_json(),
+                        })?
+                        .into(),
                     ))
                     .await;
             }
             rabbitmq::Message::EnableRaiseHands { issued_by } => {
                 self.ws
                     .send(Message::Text(
-                        NamespacedOutgoing {
+                        serde_json::to_string(&NamespacedEvent {
                             namespace: moderation::NAMESPACE,
                             timestamp,
                             payload: moderation::outgoing::Message::RaiseHandsEnabled { issued_by },
-                        }
-                        .to_json(),
+                        })?
+                        .into(),
                     ))
                     .await;
             }
@@ -1840,14 +1841,14 @@ impl Runner {
 
                 self.ws
                     .send(Message::Text(
-                        NamespacedOutgoing {
+                        serde_json::to_string(&NamespacedEvent {
                             namespace: moderation::NAMESPACE,
                             timestamp,
                             payload: moderation::outgoing::Message::RaiseHandsDisabled {
                                 issued_by,
                             },
-                        }
-                        .to_json(),
+                        })?
+                        .into(),
                     ))
                     .await;
             }
@@ -1865,7 +1866,7 @@ impl Runner {
         recipient: Option<ParticipantId>,
         message: rabbitmq::Message,
     ) {
-        let message = Namespaced {
+        let message = NamespacedCommand {
             namespace: NAMESPACE,
             payload: message,
         };
@@ -1876,8 +1877,13 @@ impl Runner {
             Cow::Borrowed(rabbitmq::room_all_routing_key())
         };
 
-        self.rabbitmq_publish(timestamp, None, &routing_key, message.to_json())
-            .await;
+        self.rabbitmq_publish(
+            timestamp,
+            None,
+            &routing_key,
+            serde_json::to_string(&message).expect("Failed to convert namespaced to json"),
+        )
+        .await;
     }
 
     /// Publish a rabbitmq message
@@ -2029,12 +2035,13 @@ impl Runner {
     async fn ws_send_control(&mut self, timestamp: Timestamp, payload: outgoing::Message) {
         self.ws
             .send(Message::Text(
-                NamespacedOutgoing {
+                serde_json::to_string(&NamespacedEvent {
                     namespace: NAMESPACE,
                     timestamp,
                     payload,
-                }
-                .to_json(),
+                })
+                .expect("Failed to convert namespaced to json")
+                .into(),
             ))
             .await;
     }