+++ crates/types/src/signaling.rs
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
+//
+// SPDX-License-Identifier: EUPL-1.2
+
+//! This module contains types that are used by the signaling communication
+//! (typically through websockets)
+
+mod namespaced;
+
+pub use namespaced::{NamespacedCommand, NamespacedEvent};